+++ app/classes/helpers/file_helpers.py
@@ -7,7 +7,6 @@ import os
 import pathlib
 import shutil
 import ssl
-import tempfile
 import time
 import urllib.request
 import zipfile
@@ -21,7 +20,6 @@ import certifi
 from app.classes.models.server_permissions import PermissionsServers
 from app.classes.helpers.cryptography_helper import CryptoHelper
 from app.classes.helpers.helpers import Helpers
-from app.classes.shared.console import Console
 from app.classes.shared.websocket_manager import WebSocketManager
 
 logger = logging.getLogger(__name__)
@@ -487,26 +485,87 @@ class FileHelpers:
         with zipfile.ZipFile(archive_location, "r") as zip_ref:
             zip_ref.extractall(destination)
 
-    def cleanup_unzip(self, temp_dir: Path, server_update: bool, new_dir: Path):
-        ignored_names = [
-            "server.properties",
-            "permissions.json",
-            "allowlist.json",
-        ]
-        # we'll iterate through the top level directory moving everything
-        # out of the temp directory and into it's final home.
-        for item in os.listdir(temp_dir):
-            # if the file is one of our ignored names we'll skip it
-            if item in ignored_names and server_update:
-                continue
-            # we handle files and dirs differently or we'll crash out.
+    def send_percentage(self, user, percent, proc_id, complete):
+        if isinstance(user, str):
+            WebSocketManager().broadcast_user(
+                user,
+                "zip_status",
+                {"id": None, "percent": percent, "complete": complete},
+            )
+        else:
+            for usr in user:
+                WebSocketManager().broadcast_user(
+                    usr,
+                    "zip_status",
+                    {"id": proc_id, "percent": percent, "complete": complete},
+                )
+
+    def should_extract(
+        self, file, base_include_path, excluded_files, server_update
+    ) -> bool:
+        """Checks a number of inclusions or exclusions against a given file to see
+        if that file should be unpacked to the target directory.
+
+        ** Base include path and excluded files should not be used in conjunction with
+        eachother.
+
+        Args:
+            file (str): file name from Path zip object namelist
+            base_include_path (str): string from root dir select that shows base path
+            like 'server_files/myserver/' (should not be used with excluded_files)
+            excluded_files (list): list of file exclusions (should not be used with base
+            include path)
+            server_update (bool): whether or not the method was called as a result
+            of a server update process.
+
+        Returns:
+            bool: Whether or not the file from the list should be included in the
+            unzipped archive.
+        """
+        if server_update and file in excluded_files:
+            return False
+
+        if not base_include_path:
+            return True
+
+        try:
+            pathlib.PurePosixPath(file).relative_to(
+                pathlib.PurePosixPath(base_include_path)
+            )
+            return True
+        except ValueError:
+            return False
+
+    def get_archive_internal_name(self, file, base_include_path) -> str:
+        """If we have a base include path we will rewrite the internal zip object path
+        to remove the /path/to/file/in/archive so we don't have nested folders when we
+        unzip. This will return the relative path to the archive to avoid nesting.
+
+        Args:
+            file (str): file from namelist from zip object
+            base_include_path (str): string from root dir select that shows base path
+            like 'server_files/myserver/'
+
+        Returns:
+            str | PurePosixPath: returns the original file name or new file name
+        """
+        if base_include_path:  # Rewrite path of zip_ref_info if we have a base include
             try:
-                self.move_item_file_or_dir(temp_dir, new_dir, item)
-            except shutil.Error as ex:
-                logger.error(f"ERROR IN ZIP IMPORT: {ex}")
+                rel = pathlib.PurePosixPath(file).relative_to(base_include_path)
+                return str(rel)
+            except ValueError:
+                logger.debug("%s is not relative to %s", file, base_include_path)
+        return str(file)
 
     def unzip_file(
-        self, zip_path, server_id, server_update: bool = False, proc_id=None
+        self,
+        zip_path,
+        destination_path,
+        server_id=None,
+        server_update: bool = False,
+        proc_id=None,
+        user_id=None,
+        base_include_path=None,
     ) -> None:
         """
         Unzips zip file at zip_path to location generated at new_dir based on zip
@@ -520,37 +579,40 @@ class FileHelpers:
         Returns: None
 
         """
-        server_users = PermissionsServers.get_server_user_list(server_id)
-        # Get directory without zipfile name
-        new_dir = pathlib.Path(zip_path).parents[0]
+        ignored_names = [
+            "server.properties",
+            "permissions.json",
+            "allowlist.json",
+        ]
+        server_users = user_id
+        if not server_users:
+            server_users = PermissionsServers.get_server_user_list(server_id)
+
         # make sure we're able to access the zip file
         if Helpers.check_file_perms(zip_path) and os.path.isfile(zip_path):
             # make sure the directory we're unzipping this to exists
-            Helpers.ensure_dir_exists(new_dir)
-            # we'll make a temporary directory to unzip this to.
-            temp_dir = tempfile.mkdtemp()
-            try:
-                with zipfile.ZipFile(zip_path, "r") as zip_ref:
-                    # we'll extract this to the temp dir using zipfile module
-                    files_list = zip_ref.namelist()
-                    for idx, file in enumerate(files_list):
-                        percent = round((idx / len(files_list)) * 100)
-                        zip_ref.extract(file, temp_dir)
-                        for user in server_users:
-                            WebSocketManager().broadcast_user(
-                                user,
-                                "zip_status",
-                                {"id": proc_id, "percent": percent, "complete": False},
-                            )
-                self.cleanup_unzip(Path(temp_dir), server_update, new_dir)
-                for user in server_users:
-                    WebSocketManager().broadcast_user(
-                        user,
-                        "zip_status",
-                        {"id": proc_id, "percent": 100, "complete": True},
-                    )
-            except Exception as ex:
-                Console.error(ex)
+            Helpers.ensure_dir_exists(destination_path)
+            with zipfile.ZipFile(zip_path, "r") as zip_ref:
+                files_list = zip_ref.namelist()
+                for idx, file in enumerate(files_list):
+                    info = zip_ref.getinfo(file)
+                    target = Path(destination_path, file).resolve()
+                    try:
+                        self.helper.validate_traversal(destination_path, target)
+                    except ValueError:
+                        self.send_percentage(server_users, 100, proc_id, True)
+                        return logger.error("Traversal detected. Dumping out.")
+                    # if the file is one of our ignored names we'll skip it
+                    if self.should_extract(
+                        file, base_include_path, ignored_names, server_update
+                    ):
+                        info.filename = self.get_archive_internal_name(
+                            file, base_include_path
+                        )
+                        zip_ref.extract(file, destination_path)
+                    percent = round((idx / len(files_list)) * 100)
+                    self.send_percentage(server_users, percent, proc_id, False)
+            self.send_percentage(server_users, 100, proc_id, True)
 
     @staticmethod
     def get_absolute_path(server_path, path) -> str:
@@ -568,16 +630,6 @@ class FileHelpers:
             path = str(Path(server_path, request_path))
 
         return str(path)
-
-    @staticmethod
-    def unzip_server(zip_path, user_id):
-        if Helpers.check_file_perms(zip_path):
-            temp_dir = tempfile.mkdtemp()
-            with zipfile.ZipFile(zip_path, "r") as zip_ref:
-                # extracts archive to temp directory
-                zip_ref.extractall(temp_dir)
-            if user_id:
-                return temp_dir
 
     @staticmethod
     def get_chunk_path_from_hash(chunk_hash: bytes, repository_location: Path) -> Path: